# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 为API生，为框架死，为debug奋斗一辈子；吃符号亏，上大小写的当，最后死在需求上   ------- 梦旋飞
  version: 1.0.0
  title: 开发平台
basePath: /ld_dlj_interface
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /industryClass/getData:
    post:
      tags:
        - 行业分类模块接口
      summary: 基础数据：当前用户数，同比，环比，用户量，装机容量，电量，产能利用率
      produces:
        - application/json
      parameters:
        - name: industryId
          description: 行业id
          in: query
          type: string
          required: true
        - name: timeType
          description:  查询时间类型 1按年，2按月。
          in: query
          type: string
          required: true

      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /industryClass/getDisData:
    post:
      tags:
        - 行业分类模块接口
      summary: 折现图数据
      produces:
        - application/json
      parameters:
        - name: industryId
          description: 行业id
          in: query
          type: string
          required: true
        - name: type
          description: 类型 1.用户量 2.装机容量 3.电量 4.产能利用率
          in: query
          type: string
          required: true
        - name: timeType
          description: 查询时间类型 1按年，2按月。
          in: query
          type: string
          required: true

      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /industryClass/getLastYearEleCon:
    post:
      tags:
        - 行业分类模块接口
      summary: 上年度居民用电比例
      produces:
        - application/json
      parameters:
        - name: startTime
          description: 开始时间 YYYY 数据从2017年开始。如果查询的是2017年，那么开始时间和结束时间就是2017-2018
          in: query
          type: string
          required: true
        - name: endTime
          description: 结束时间 YYYY
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string